requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
idna==2.6
pycparser==2.18
pyOpenSSL==17.5.0
requests==2.18.4
six==1.11.0
tornado==5.0
urllib3==1.22
wheel==0.24.0

file:../botie
